requirements.txt
Django>=5.0.5
psycopg2-binary>=2.9.9
django-compressor>=4.4
pillow>=10.3.0
requests>=2.31.0
python-dotenv>=1.0.1
folium>=0.16.0